+++ Modules/Core/QuadEdgeMesh/test/itkQuadEdgeMeshEulerOperatorDeleteCenterVertexTest.cxx
@@ -16,6 +16,8 @@
  *
  *=========================================================================*/
 
+#include <cstddef>
+
 #include "itkQuadEdgeMeshEulerOperatorDeleteCenterVertexFunction.h"
 #include "itkQuadEdgeMeshEulerOperatorCreateCenterVertexFunction.h"
 #include "itkQuadEdgeMeshEulerOperatorsTestHelper.h"
@@ -131,7 +133,7 @@ itkQuadEdgeMeshEulerOperatorDeleteCenterVertexTest(int itkNotUsed(argc), char * 
     {
       poly = new QEPolygonCellType(3);
       cellpointer.TakeOwnership(poly);
-      cellpointer->SetPointId(0, specialCells[3 * i]);
+      cellpointer->SetPointId(0, specialCells[static_cast<ptrdiff_t>(3 * i)]);
       cellpointer->SetPointId(1, specialCells[3 * i + 1]);
       cellpointer->SetPointId(2, specialCells[3 * i + 2]);
       specialmesh->SetCell(i, cellpointer);
