+++ Modules/Core/QuadEdgeMesh/test/itkQuadEdgeMeshFrontIteratorTest.cxx
@@ -28,6 +28,8 @@
  * primal and dual version of the front iterator.
  */
 
+#include <cstddef>
+
 #include "itkQuadEdgeMesh.h"
 
 int
@@ -178,7 +180,7 @@ itkQuadEdgeMeshFrontIteratorTest(int, char *[])
   {
     poly = new QEPolygonCellType(3);
     cellpointer.TakeOwnership(poly);
-    cellpointer->SetPointId(0, simpleSquareCells[3 * i]);
+    cellpointer->SetPointId(0, simpleSquareCells[static_cast<ptrdiff_t>(3 * i)]);
     cellpointer->SetPointId(1, simpleSquareCells[3 * i + 1]);
     cellpointer->SetPointId(2, simpleSquareCells[3 * i + 2]);
     mesh->SetCell(i, cellpointer);
